Refactor routing according to the OAuth tutorial before getting into DB setup.
