+++ server/src/routes/index.ts
@@ -1,0 +1,10 @@
+import { Router } from "express";
+import authRoutes from "./auth";
+import userRoutes from "./user";
+
+const router = Router();
+
+router.use("/auth", authRoutes);
+router.use("/user", userRoutes);
+
+export default router;
